Completely refactored this module to support a more middleware-type of system. Implemented everything, but it's not completely tested. I've been finding issues with the implementation while trying to use it, but it's mostly solid and should work when I need to add in the database middleware.
It seems odd to think that all this work has been done, but I haven't even started on the database piece. The interfaces in vsql are solid, that's what has enabled all of this work to flow without much re-work of the vsql module.

The worst part of this process has been duplicating the middleware handler logic due to the way Go works with "inheritance". But I believe I've avoided any major pitfalls.

Ideally, one of these days, I'll add in a code generator for the contexts and the middlewares. However, this will do for now.

I've tested the non-nested engine and statements. Tests that remain to be built: transactions, nested transactions